changes links for failshot and video.flv to be relative to the final.html.
The previous implentation was expecting Jenkins workspaces to be in /ws/ which they are not for all job types. This does mean that on jenkins we would need to archive failshots and videos so that they are are available on the relative path.

untested think these should be in too